break down employee summary view page
