+++ react-app/src/index.jsx
@@ -3,7 +3,7 @@ import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
 import App from "./App.jsx";
 import store from "./store/store.js";
-import EmployeeSummaryView from "./components/employeeSummaryView.jsx";
+import EmployeeSummaryView from "./pages/EmployeeSummaryView.jsx";
 import VisaStatusManagementPage from "./components/VisaStatusHr.jsx";
 import UserVisaPage from "./pages/UserVisaPage.jsx";
 import Modal from "react-modal";
